# Implement number checks in C