I was hoping for more of a speed-up, especially given how much effort it is to write these things in C. I guess the R versions are already quite fast to begin with. Will this be enough to fix the perf issues @DavisVaughan?

```r
bench::mark(
  new = check_number_whole(1),
  old = check_number_whole_old(1),
  iterations = 10000
)
#> # A tibble: 2 × 13
#>   express…¹    min median itr/s…² mem_a…³ gc/se…⁴ n_itr  n_gc total…⁵
#>   <bch:exp> <bch:> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t>
#> 1 new       1.07µs 1.44µs 675435.      0B     0   10000     0  14.8ms
#> 2 old       3.57µs 3.81µs 254095.      0B    25.4  9999     1  39.4ms

bench::mark(
  new = check_number_decimal(1.5),
  old = check_number_decimal_old(1.5),
  iterations = 10000
)
#> # A tibble: 2 × 13
#>   express…¹    min median itr/s…² mem_a…³ gc/se…⁴ n_itr  n_gc total…⁵
#>   <bch:exp> <bch:> <bch:>   <dbl> <bch:b>   <dbl> <int> <dbl> <bch:t>
#> 1 new       1.19µs 1.64µs 572827.      0B       0 10000     0  17.5ms
#> 2 old       3.03µs 3.53µs 259783.      0B       0 10000     0  38.5ms
```

FYI @hadley