@lionel- 2023-02-15
oops benchmarks above are with `-g -O0`, the `-O3` benchmarks are better:

```r
bench::mark(
  new = check_number_whole(1),
  old = check_number_whole_old(1),
  iterations = 100000
)
#> # A tibble: 2 × 13
#>   expression      min   median `itr/sec` mem_al…¹ gc/se…² n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:by>   <dbl> <int> <dbl>
#> 1 new        573.93ns 737.84ns  1238480.       0B    61.9 99995     5
#> 2 old          2.67µs   3.03µs   321286.       0B    51.4 99984    16

bench::mark(
  new = check_number_decimal(1.5),
  old = check_number_decimal_old(1.5),
  iterations = 10000
)
#> # A tibble: 2 × 13
#>   expression      min   median `itr/sec` mem_al…¹ gc/se…² n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:by>   <dbl> <int> <dbl>
#> 1 new        614.91ns 738.07ns  1221700.       0B    61.1 99995     5
#> 2 old          2.34µs   2.71µs   359769.       0B    57.6 99984    16

````