@DavisVaughan 2023-02-15
I think it will help in `lead()` and `lag()` for sure. If we didn't make this change then after I add in faster vctrs assertions this would probably show up as one of the larger points of slowness

```r
library(dplyr)

x <- 1:5 + 0L

bench::mark(lead(x), iterations = 100000)

# Main (dev dplyr, dev rlang, dev vctrs)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 lead(x)      32.1µs   43.2µs    22647.     211KB     15.6

# With this PR and updated standalone file
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 lead(x)      21.6µs   28.7µs    34649.     131KB     15.3

# simulate many groups
bench::mark(for(i in 1:10000) lead(x), iterations = 10)

# Main (dev dplyr, dev rlang, dev vctrs)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 for (i in 1:10000) lead(x)    350ms    358ms      2.73    2.56MB     20.8

# With this PR and updated standalone file
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 for (i in 1:10000) lead(x)    226ms    264ms      3.82    2.56MB     19.5
```