@DavisVaughan 2023-02-15
@lionel- I think your M1 runs R code much faster than my Intel. So the switch to C seems more noticeable here

```r
x <- 1L

bench::mark(check_number_whole(x), iterations = 500000)

# Main
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 check_number_whole(x)   7.01µs   9.17µs   106014.    2.08KB     17.4

# This PR
#> # A tibble: 1 × 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 check_number_whole(x)   1.99µs   2.34µs   402779.        0B     12.1
```